+++ src/generators/generateAssistantStubs.ts
@@ -17,12 +17,23 @@ See the License for the specific language governing permissions and
 import * as vscode from 'vscode';
 import * as fs from 'fs';
 import * as path from 'path';
+<<<<<<< HEAD:src/generators/generateAssistantStubs.ts
 import { ArtificialApollo, Assistant, AssistantTypeInfo } from '../providers/apolloProvider';
 import { OutputLog } from '../providers/outputLogProvider';
 import { snakeCase } from 'lodash';
 import { AssistantByLabTreeView } from '../views/assistantTreeView';
 import { AssistantSignature, BuildAssistantSignatures } from '../parsers/parseAssistantSignatures';
 import * as _ from 'lodash';
+=======
+import { pathExists } from '../utils';
+import { ArtificialApollo, Assistant, AssistantTypeInfo } from '../providers/apolloProvider';
+import { OutputLog } from '../providers/outputLogProvider';
+import { snakeCase, camelCase } from 'lodash';
+import { BuildPythonSignatures } from '../parsers/parsePythonSignatures';
+import { AssistantByLabTreeView } from '../views/assistantTreeView';
+import { BuildAssistantSignatures } from '../parsers/parseAssistantSignatures';
+import _ = require('lodash');
+>>>>>>> main:src/generators/generateActionStubs.ts
 
 export class GenerateAssistantStubs {
   outputChannel = OutputLog.getInstance();
@@ -49,15 +60,29 @@ export class GenerateAssistantStubs {
     if (!response?.assistants) {
       return;
     }
-
-    for (const sig of response?.assistants) {
-      let allParamsSorted = this.getAllParamsSorted(sig, stubs);
-      pythonContent += `@assistant('${sig.id}')\n`;
-      pythonContent += this.buildAssistantParmDec(sig, allParamsSorted);
-      pythonContent += `async def assistant_${snakeCase(sig.name)}(\n`;
-      pythonContent += this.buildAssistantParams(sig, allParamsSorted);
-      pythonContent += `) -> None:\n`;
-      pythonContent += `    pass\n\n\n`;
+    const labs = await client.queryLabs();
+    if (!labs) {
+      return;
+    }
+    for (const lab of labs.labs) {
+      pythonContent += `class ${lab.name.charAt(0).toUpperCase() + camelCase(lab.name).slice(1)}Assistants:\n`;
+      let labContainsAssistants = false;
+      for (const sig of response.assistants) {
+        if (lab.id === sig.constraint.labId) {
+          labContainsAssistants = true;
+          let allParamsSorted = this.getAllParamsSorted(sig, stubs);
+          pythonContent += `\t@staticmethod\n`;
+          pythonContent += `\t@assistant('${sig.id}')\n`;
+          pythonContent += this.buildAssistantParmDec(sig, allParamsSorted);
+          pythonContent += `\tasync def assistant_${snakeCase(sig.name)}(\n`;
+          pythonContent += this.buildAssistantParams(sig, allParamsSorted);
+          pythonContent += `\t) -> None:\n`;
+          pythonContent += `\t\tpass\n\n\n`;
+        }
+      }
+      if (!labContainsAssistants) {
+        pythonContent += '\tpass\n\n\n';
+      }
     }
 
     fs.writeFile(fullPath, pythonContent, (err) => {
@@ -67,20 +92,30 @@ export class GenerateAssistantStubs {
     });
   }
 
+  //TODO: Once we have fully updated all NS to use asst params with indices, this can be removed
   private getAllParamsSorted(apolloSignature: Assistant, stubs: AssistantSignature[]): string[] {
     let allParamsSorted: string[] = [];
-
     let stubParams = [];
     let alabParams = [];
+
+    // Check if asst params have indices set
+    let usingIndices = false;
+    for (let param of apolloSignature.parameters) {
+      alabParams.push(param.typeInfo.name);
+      if (param.index > 0) {
+        usingIndices = true;
+      }
+    }
+    if (usingIndices) {
+      return alabParams;
+    }
+    // If no explicit order set, fallback to using stubs as the order if they exist
     if (stubs) {
       const stubSignature = stubs.find((element) => element.actionId === apolloSignature.id);
       if (stubSignature) {
         for (let param of stubSignature?.parameters) {
           stubParams.push(param.assistantName);
         }
-      }
-      for (let param of apolloSignature.parameters) {
-        alabParams.push(param.typeInfo.name);
       }
       const sortedParams = _.intersection(stubParams, alabParams);
       allParamsSorted = _.concat(sortedParams, _.difference(alabParams, sortedParams));
@@ -95,7 +130,7 @@ export class GenerateAssistantStubs {
     for (const param of allParamsSorted) {
       const paramSig = sig.parameters.find((element) => element.typeInfo.name === param);
       if (paramSig) {
-        returnString += `    arg_${snakeCase(paramSig.typeInfo.name)}: ${this.convertToPythonType(
+        returnString += `\t\targ_${snakeCase(paramSig.typeInfo.name)}: ${this.convertToPythonType(
           paramSig.typeInfo
         )},\n`;
       }
@@ -136,7 +171,7 @@ export class GenerateAssistantStubs {
     for (const param of allParamsSorted) {
       const paramSig = sig.parameters.find((element) => element.typeInfo.name === param);
       if (paramSig) {
-        returnString += `@parameter('arg_${snakeCase(paramSig.typeInfo.name)}', action_parameter_name='${
+        returnString += `\t@parameter('arg_${snakeCase(paramSig.typeInfo.name)}', action_parameter_name='${
           paramSig.typeInfo.name
         }')\n`;
       }