+++ src/views/assistantTreeView.ts
@@ -17,15 +17,13 @@ See the License for the specific language governing permissions and
 import * as vscode from 'vscode';
 import * as path from 'path';
 import { ArtificialApollo, AssistantReply, Assistant, AssistantTypeInfo } from '../providers/apolloProvider';
-import { BuildAssistantSignatures, AssistantSignature } from '../parsers/parseAssistantSignatures';
+import { BuildAssistantSignatures } from '../parsers/parseAssistantSignatures';
 import { LabTreeElement } from './loadConfigTreeView';
 import * as _ from 'lodash';
+import { camelCase } from 'lodash';
 
 type TreeElement = LabTreeElement | AssistantTreeElement;
-interface Error {
-  code: number;
-  error: string;
-}
+
 export class AssistantByLabTreeView
   implements vscode.TreeDataProvider<TreeElement>, vscode.TreeDragAndDropController<TreeElement>
 {
@@ -137,9 +135,9 @@ export class AssistantByLabTreeView
         if (found.constraint.labId === element.labId) {
           const result = this.validParams(sig, found);
           if (result.code === 0) {
-            treeElements.push(new AssistantTreeElement(sig.name, element.labId, sig));
+            treeElements.push(new AssistantTreeElement(sig.name, element.labId, element.label, sig));
           } else {
-            treeElements.push(new AssistantTreeElementError(sig.name, element.labId, sig, result.error));
+            treeElements.push(new AssistantTreeElementError(sig.name, element.labId, element.label, sig, result.error));
           }
         }
       } else if (element.label === 'UNKNOWN') {
@@ -147,16 +145,37 @@ export class AssistantByLabTreeView
           new AssistantTreeElementError(
             sig.name,
             element.labId,
+            element.label,
             sig,
             'Assistant does not match a known lab in Artificial Cloud'
           )
         );
       }
     }
+    // Find cloud assistants with no stubs
+    if (element.label !== 'UNKNOWN') {
+      for (const assistant of response.assistants) {
+        const found = treeElements.find((ele) => ele.functionSignature.actionId === assistant.id);
+        if (!found) {
+          if (assistant.constraint.labId === element.labId) {
+            const blankSignature: AssistantSignature = { actionId: '', parameters: [], name: '' };
+            treeElements.push(
+              new AssistantTreeElementError(
+                assistant.name,
+                element.labId,
+                element.label,
+                blankSignature,
+                'No Stub for ALab Assistant'
+              )
+            );
+          }
+        }
+      }
+    }
     return treeElements;
   }
 
-  private validParams(stubSignature: AssistantSignature, assistant: Assistant): Error {
+  private validParams(stubSignature: AssistantSignature, assistant: Assistant): AssistantTypeError {
     const stubNames: string[] = [];
     const assistantParamNames: string[] = [];
     for (const param of stubSignature.parameters) {
@@ -236,6 +255,7 @@ export class AssistantTreeElement extends vscode.TreeItem {
   constructor(
     public readonly label: string,
     public readonly labId: string,
+    public readonly labName: string,
     public readonly functionSignature: AssistantSignature
   ) {
     super(label, vscode.TreeItemCollapsibleState.None);
@@ -243,7 +263,8 @@ export class AssistantTreeElement extends vscode.TreeItem {
     this.functionSignature = functionSignature;
     this.contextValue = 'ASSISTANT';
   }
-  resourceUri = vscode.Uri.parse('artificial/assistantByLab/assistant/' + this.label);
+  labClassName = this.labName.charAt(0).toUpperCase() + camelCase(this.labName).slice(1) + 'Assistants';
+  resourceUri = vscode.Uri.parse('artificial/assistantByLab/assistant/' + this.labClassName + '/' + this.label);
   type = 'assistant';
   iconPath = {
     light: path.join(__filename, '..', '..', 'resources', 'light', 'assistants.svg'),
@@ -254,6 +275,7 @@ export class AssistantTreeElementError extends vscode.TreeItem {
   constructor(
     public readonly label: string,
     public readonly labId: string,
+    public readonly labName: string,
     public readonly functionSignature: AssistantSignature,
     public readonly tooltip: string
   ) {
@@ -263,6 +285,7 @@ export class AssistantTreeElementError extends vscode.TreeItem {
     this.contextValue = 'ASSISTANT';
     this.description = tooltip;
   }
+  labClassName = this.labName.charAt(0).toUpperCase() + camelCase(this.labName).slice(1) + 'Assistants';
   resourceUri = vscode.Uri.parse('artificial/typeError/assistantByLab/assistant/' + this.functionSignature.name);
   type = 'assistant';
   iconPath = {